requirements.txt
openai
tensorflow==2.15.0
streamlit
pillow
tf-bodypix
